% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cov.sc}
\alias{cov.sc}
\title{Coverage for putative heterozygous variants inferred from the ExAC database
for cells from MGH31 from Patel. et al.}
\format{An object of class \code{matrix} with 5587 rows and 75 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/24925914}
}
\usage{
cov.sc
}
\description{
Coverage for putative heterozygous variants inferred from the ExAC database
for cells from MGH31 from Patel. et al.
}
\keyword{datasets}
