% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonal_deletion_detection.R
\name{getHetRate}
\alias{getHetRate}
\title{Helper function to get het rate from allele counts at a list of common variant
SNPs}
\usage{
getHetRate(snpSummary, t = 1e-08, alt = FALSE)
}
\arguments{
\item{snpSummary}{data frame of common variant SNPs and their allele counts and
coverage}

\item{t}{threshold for hypothesis testing for heterozygosity, default 1e-8
(Illumina sequencing error rate)}

\item{alt}{Boolean of whether or not to use alternate het filter}
}
\value{
hetRate percentage of heterozygous SNPs in the input data frame
}
\description{
Helper function to get het rate from allele counts at a list of common variant
SNPs
}
\examples{
\dontrun{
}

}
