% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gexp}
\alias{gexp}
\title{Gene expression for cells from MGH31 from Patel et al.}
\format{An object of class \code{matrix} with 6082 rows and 75 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/24925914}
}
\usage{
gexp
}
\description{
Gene expression for cells from MGH31 from Patel et al.
}
\keyword{datasets}
