% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeybadger.R
\docType{class}
\name{honeybadger-class}
\alias{honeybadger-class}
\alias{honeybadger}
\title{A Reference Class to represent single-cell RNA-seq data for HoneyBADGER analysis}
\description{
A Reference Class to represent single-cell RNA-seq data for HoneyBADGER analysis
}

