% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoneyBADGER.R
\name{HoneyBADGER_calcAlleleCnvProb}
\alias{HoneyBADGER_calcAlleleCnvProb}
\title{Calculate posterior probability of CNVs using allele data}
\arguments{
\item{r.sub}{Optional matrix of alt allele count in single cells. If not provided, internal r.sc matrix is used.}

\item{n.sub}{Optional matrix of site coverage count in single cells. If not provided, internal n.sc matrix is used.}

\item{l.sub}{Optional vector of alt allele count in pooled single cells or bulk. If not provided, internal l vector is used.}

\item{n.bulk.sub}{Optional vector of site coverage count in pooled single cells or bulk. If not provided, internal n.bulk vector is used.}

\item{region}{GenomicRanges region of interest such as expected CNV boundaries.}

\item{filter}{Boolean for whether to filter out SNP sites with no coverage. (default: TRUE)}

\item{pe}{Effective error rate to capture error from sequencing, etc. (default: 0.01)}

\item{mono}{Rate of mono-allelic expression. (default: 0.7)}

\item{n.iter}{Number of iterations in MCMC. (default: 1000)}

\item{quiet}{Boolean of whether to suppress progress bar. (default: TRUE)}

\item{verbose}{Verbosity(default: FALSE)}
}
\description{
Calculate posterior probability of CNVs using allele data
}
