% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeybadger.R
\name{honeybadger_calcCombCnvProb

mu0'  = mu0,
sigma0' = sigma0
t' = mu0/2}
\alias{honeybadger_calcCombCnvProb

mu0'  = mu0,
sigma0' = sigma0
t' = mu0/2}
\title{Calculate posterior probability of CNVs using normalized expression data and allele data}
\description{
Calculate posterior probability of CNVs using normalized expression data and allele data
}
