% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeybadger.R
\name{honeybadger_calcCombCnvProb}
\alias{honeybadger_calcCombCnvProb}
\title{Calculate posterior probability of CNVs using normalized expression data and allele data}
\description{
Calculate posterior probability of CNVs using normalized expression data and allele data
}
