% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoneyBADGER.R
\name{HoneyBADGER_calcGexpCnvBoundaries}
\alias{HoneyBADGER_calcGexpCnvBoundaries}
\title{Recursive HMM to identify CNV boundaries using normalized gene expression data}
\description{
Recursive HMM to identify CNV boundaries using normalized gene expression data
}
