% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeybadger.R
\name{honeybadger_calcGexpCnvProb}
\alias{honeybadger_calcGexpCnvProb}
\title{Calculate posterior probability of CNVs using normalized expression data}
\arguments{
\item{gexp.norm.sub}{Optional normalized gene expression matrix. If not provided, internal normalized gene expression matrix is used.}

\item{m}{Expected mean deviation due to copy number change (default: 0.15)}

\item{region}{GenomicRanges region of interest such as expected CNV boundaries}

\item{quiet}{Boolean for whether to suppress progress display}
}
\description{
Calculate posterior probability of CNVs using normalized expression data
}
