% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoneyBADGER.R, R/HoneyBADGER_allele.R
\name{HoneyBADGER_plotAlleleProfile}
\alias{HoneyBADGER_plotAlleleProfile}
\alias{HoneyBADGER_plotAlleleProfile}
\alias{plotAlleleProfile}
\title{Plot allele profile}
\usage{
plotAlleleProfile(r.maf, n.sc, l.maf, n.bulk, region = NULL,
  chrs = paste0("chr", c(1:22)), widths = NULL, cellOrder = NULL,
  filter = FALSE, max.ps = 3)
}
\arguments{
\item{region}{Limit plotting to particular GenomicRanges regions}

\item{chrs}{Limit plotting to select chromosomes. Default autosomes only. (default: paste0('chr', c(1:22)))}

\item{widths}{Widths of chromosomes in plot. If 'set' will depend on number of SNPs in region. Else will be equal.}

\item{cellOrder}{Order of cells. Otherwise will be same order as in input matrix}

\item{filter}{Remove sites with no coverage}

\item{max.ps}{Maximum point size for plot}

\item{r.sub}{SNP lesser allele count matrix for single cells. If NULL, object's r.maf will be used}

\item{n.sc.sub}{SNP coverage count matrix for single cells. If NULL, object's n.sc will be used}

\item{l.sub}{SNP lesser allele count matrix for bulk refernece. If NULL, object's l.maf will be used}

\item{n.bulk.sub}{SNP coverage count matrix for bulk refernece. If NULL, object's n.bulk will be used}

\item{setWidths}{Set widths on chromosomes based on SNP density. Otherwise will be uniform.}

\item{returnPlot}{Whether to return ggplot object}

\item{r.sub}{SNP lesser allele count matrix for single cells. If NULL, object's r.maf will be used}

\item{n.sc.sub}{SNP coverage count matrix for single cells. If NULL, object's n.sc will be used}

\item{l.sub}{SNP lesser allele count matrix for bulk refernece. If NULL, object's l.maf will be used}

\item{n.bulk.sub}{SNP coverage count matrix for bulk refernece. If NULL, object's n.bulk will be used}

\item{region}{Limit plotting to particular GenomicRanges regions}

\item{chrs}{Limit plotting to select chromosomes. Default autosomes only. (default: paste0('chr', c(1:22)))}

\item{cellOrder}{Order of cells. If 'set' will be automatically ordered by clustering. Else will be same order as input.}

\item{filter}{Remove sites with no coverage}

\item{max.ps}{Maximum point size for plot.}
}
\description{
Plot allele profile

Plot allele profile
}
\examples{
data(r)
data(cov.sc)
allele.mats <- setAlleleMats(r, cov.sc)
plotAlleleProfile(allele.mats$r.maf, allele.mats$n.sc, allele.mats$l.maf, allele.mats$n.bulk, widths=c(249250621, 243199373, 198022430, 191154276, 180915260, 171115067, 159138663, 146364022, 141213431, 135534747, 135006516, 133851895, 115169878, 107349540, 102531392, 90354753, 81195210, 78077248, 59128983, 63025520, 51304566, 48129895)/1e7) 

}
