% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeybadger.R
\name{honeybadger_retestIdentifiedCnvs}
\alias{honeybadger_retestIdentifiedCnvs}
\title{Calculate posterior probability of CNVs for regions identified by the recursive HMM approach}
\description{
Calculate posterior probability of CNVs for regions identified by the recursive HMM approach
}
