% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoneyBADGER.R
\name{HoneyBADGER_setGeneFactors}
\alias{HoneyBADGER_setGeneFactors}
\title{Maps snps to genes}
\arguments{
\item{txdb}{TxDb object ex. TxDb.Hsapiens.UCSC.hg19.knownGene}

\item{fill}{SNPs mapping to genes not annotated in txdb will be given unique IDs}

\item{verbose}{Verbosity}
}
\description{
Maps snps to genes
}
