% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoneyBADGER.R
\name{HoneyBADGER_setGexpDev}
\alias{HoneyBADGER_setGexpDev}
\title{Set needed absolute gene expression deviance to be able to distinguish neutral from amplified or deletion regions}
\arguments{
\item{alpha}{Alpha level (default: 0.05)}

\item{n}{Number of repeats for estimating parameter (default: 100)}

\item{seed}{Random seed}

\item{plot}{Plotting}

\item{verbose}{Verbosity}
}
\description{
Set needed absolute gene expression deviance to be able to distinguish neutral from amplified or deletion regions
}
