% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoneyBADGER.R
\name{HoneyBADGER_setMvFit}
\alias{HoneyBADGER_setMvFit}
\title{Model expected gene expression variance as a function of number of genes}
\arguments{
\item{num.genes}{Number of random genes sampled (default: seq(5, 100, by=5))}

\item{rep}{Number of repeats/resampling (default: 50)}

\item{plot}{Whether to plot (default: FALSE)}

\item{verbose}{Verbosity (default: TRUE)}
}
\description{
Model expected gene expression variance as a function of number of genes
}
