% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoneyBADGER.R
\name{HoneyBADGER_summarizeResults}
\alias{HoneyBADGER_summarizeResults}
\title{Summarize results}
\arguments{
\item{geneBased}{Boolean of whether to summarize gene-based results}

\item{alleleBased}{Boolean of whether to summarize allele-based results}

\item{min.num.cells}{CNV must be present in this minimum number of cells}

\item{t}{Minimum probability threshold for call}
}
\description{
Summarize results
}
