% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoneyBADGER.R
\name{HoneyBADGER_visualizeResults}
\alias{HoneyBADGER_visualizeResults}
\title{Plot posterior probability as heatmap}
\arguments{
\item{geneBased}{Boolean whether to use gene-based results summary}

\item{alleleBased}{Boolean whether to use allele-based results summary}

\item{hc}{hclust object for cells}

\item{vc}{hclust object for CNVs}

\item{power}{Parameter to tweak clustering by weighing posterior probabilities}

\item{details}{Boolean whether to return details}

\item{...}{Additional parameters to pass to heatmap}
}
\description{
Plot posterior probability as heatmap
}
