% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref}
\alias{ref}
\title{Gene expression reference using normal cells from MGH31 from Patel et al.}
\format{An object of class \code{numeric} of length 6082.}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/24925914}
}
\usage{
ref
}
\description{
Gene expression reference using normal cells from MGH31 from Patel et al.
}
\keyword{datasets}
