% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoneyBADGER.R
\name{HoneyBADGER_calcAlleleCnvBoundaries}
\alias{HoneyBADGER_calcAlleleCnvBoundaries}
\title{Recursive HMM to identify CNV boundaries using allele data}
\description{
Recursive HMM to identify CNV boundaries using allele data
}
