% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoneyBADGER.R
\name{HoneyBADGER_calcCombCnvProb}
\alias{HoneyBADGER_calcCombCnvProb}
\title{Calculate posterior probability of CNVs using normalized expression data and allele data}
\description{
Calculate posterior probability of CNVs using normalized expression data and allele data
}
