% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoneyBADGER.R
\name{HoneyBADGER_calcGexpCnvProb}
\alias{HoneyBADGER_calcGexpCnvProb}
\title{Calculate posterior probability of CNVs using normalized expression data}
\arguments{
\item{gexp.norm.sub}{Optional normalized gene expression matrix. If not provided, internal normalized gene expression matrix is used.}

\item{m}{Expected mean deviation due to copy number change (default: 0.15)}

\item{region}{Optional GenomicRanges region of interest such as expected CNV boundaries. (default: NULL)}

\item{quiet}{Boolean for whether to suppress progress display (default: TRUE)}

\item{verbose}{Verbosity (default: FALSE)}
}
\description{
Calculate posterior probability of CNVs using normalized expression data
}
