% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoneyBADGER.R
\name{HoneyBADGER_plotAlleleProfile}
\alias{HoneyBADGER_plotAlleleProfile}
\title{Plot allele profile}
\arguments{
\item{r.sub}{SNP lesser allele count matrix for single cells. If NULL, object's r.maf will be used}

\item{n.sc.sub}{SNP coverage count matrix for single cells. If NULL, object's n.sc will be used}

\item{l.sub}{SNP lesser allele count matrix for bulk refernece. If NULL, object's l.maf will be used}

\item{n.bulk.sub}{SNP coverage count matrix for bulk refernece. If NULL, object's n.bulk will be used}

\item{region}{Limit plotting to particular GenomicRanges regions}

\item{chrs}{Limit plotting to select chromosomes. Default autosomes only. (default: paste0('chr', c(1:22)))}

\item{setWidths}{Set widths on chromosomes based on SNP density. Otherwise will be uniform.}

\item{cellOrder}{Order of cells. Otherwise will be same order as in input matrix}

\item{filter}{Remove sites with no coverage}

\item{returnPlot}{Whether to return ggplot object}

\item{max.ps}{Maximum point size for plot}
}
\description{
Plot allele profile
}
