% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeybadger.R
\name{honeybadger_setAlleleMats}
\alias{honeybadger_setAlleleMats}
\title{Set allele count matrices, creates in-silico pooled single cells as bulk reference if none provided}
\description{
Set allele count matrices, creates in-silico pooled single cells as bulk reference if none provided
}
